+++ zigpy/event/__init__.py
@@ -1,0 +1,4 @@
+from .diagnostic_statistic import DiagnosticStatistic, DiagnosticStatisticChangeEvent
+from .event_base import EventBase
+
+__all__ = ["DiagnosticStatistic", "DiagnosticStatisticChangeEvent", "EventBase"]
